requirements.txt
amqp==1.4.9
anyjson==0.3.3
asgiref==3.6.0
billiard==3.3.0.23
celery==3.1.26.post2
Django==4.1.7
django-celery==3.2.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
kombu==3.0.37
PyJWT==2.6.0
pytz==2022.7.1
sqlparse==0.4.3
